Plan of our project
We present to you a planning file for an innovative project that aims to develop an intelligent recommendation system for educational courses, based on the agile methodology to achieve the maximum degree of flexibility and adaptation to user requirements. The project relies on the K-means algorithm to provide personalized recommendations on platforms such as Coursera, Edraak, and other platforms. Arabic.